Quick Draw, Full Metal Body, experimental Curious Medicine, general cleanup
Phased out Teravolt, Turboblaze and Filter to squeeze in Quick Draw, Curious Medicine and Full Metal Body.

Full Metal Body was commented out and replaced with Clear Body in vanilla CFRU, but it has a minor difference in that it cannot be ignored by abilities like Mold Breaker. I reimplemented it with this in mind. The aforementioned abilities being replaced do NOT have any differences from the other abilities that are being implemented over them.

The Quick Draw code is by Greenphx9.

I have a non-working, untested Curious Medicine I commented out. It would not function whatsoever if uncommented, but I'm leaving it as a little mental bookmark for when I come back to it. Hoping to get that working tomorrow.