Work on ability/move table updates & refreshes
I've been reworking my code structure based on the CFRU dev branch and greenphx's branch, but I've been kinda... piecemeal copypasting it all as needed, and I'm starting to hit a bit of a situation where I have Frankenstein code.

There's gotta be a better way to do this than I've been doing; it might be better to simply migrate entirely to the dev branch. Consider this something of a sanity check commit so I have something to fall back on